requirements.txt
alabaster==0.7.12
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coverage==4.5.3
dj-database-url==0.5.0
Django==2.0.13
django-bootstrap-modal-forms==1.4.1
django-heroku==0.3.1
django-nose==1.4.6
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
docutils==0.14
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
nose==1.3.7
packaging==19.0
psycopg2==2.8.2
Pygments==2.3.1
pyparsing==2.4.0
pytz==2018.9
requests==2.21.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
urllib3==1.24.2
whitenoise==4.1.2
m2r==0.2.1
django-docs==0.3.1
